+++ src/app/auth/logout/components/LogoutComponent.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import { useRouter } from 'next/navigation';
+import { useEffect } from 'react';
+
+import { patchLogout } from '@/apis';
+import useUserState from '@/context/userState';
+
+export default function LogoutComponent() {
+  const { token, clearUser } = useUserState();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (token) {
+      patchLogout(token.accessToken, token.refreshToken).then(() => {
+        alert('로그아웃 되었습니다');
+      });
+    }
+    clearUser();
+    localStorage.removeItem('accesstoken');
+    router.back();
+  });
+  return <div></div>;
+}
