+++ src/app/auth/logout/components/index.ts
@@ -1,0 +1,1 @@
+export { default as LogoutComponent } from './LogoutComponent';
